publication/requirements.txt
numpy
scipy
matplotlib=1.5.1  # newer version of matplotlib and Pillow cannot handle properly alpha channel for eps output
pillow=3.2.0
pandas
statsmodels
networkx
h5py
pyyaml
brian2tools
tqdm